Relocate test to sandbox-friendly file
